reporting api

+++ backend/OpenWright.Api/Reporting/Payloads/CreateRunCasePayload.cs
@@ -8,7 +8,7 @@ public class CreateRunCasePayload
     public required Guid Id { get; init; }
     public required string Title { get; init; }
     public string[]? Tags { get; init; }
-    public long? Timeout { get; init; }
+    public decimal? Timeout { get; init; }
     public int? Retries { get; init; }
     public Annotation[]? Annotations { get; init; }
     public TestStatus? ExpectedStatus { get; init; }
